remove socket from app.js to socket controller
