+++ README.md
@@ -1,1 +1,2 @@
-# progress
+# Myprogress
+A socia media platform that aims to help an individual to record their pro
